Checker: File reorganization
