+++ pisek/jobs/task_pipeline.py
@@ -36,7 +36,7 @@ from pisek.task_jobs.generator.manager import (
     RunGenerator,
     TestcaseInfoMixin,
 )
-from pisek.task_jobs.judge import JudgeManager
+from pisek.task_jobs.checker.checker_manager import JudgeManager
 from pisek.task_jobs.builder.build import BuildManager
 from pisek.task_jobs.solution.manager import SolutionManager
 from pisek.task_jobs.testing_log import CreateTestingLog
